use app.response function instead of app.response.ask
