+++ functions/src/actions/playback-nearly-finished.js
@@ -1,12 +1,11 @@
-const {debug} = require('../utils/logger')('ia:actions:playback-stopped');
+const {debug} = require('../utils/logger')('ia:actions:playback-nearly-finished');
 
 function handler (app) {
-  // TODO: log
   debug('token', app.params.getByName('token'));
 }
 
 /**
- * handle Alexa AudioPlayer.PlaybackStopped intent
+ * handle Alexa AudioPlayer.PlaybackNearlyFinished intent
  * @type {{handler: handler}}
  */
 module.exports = {
