+++ functions/src/platform/alexa/parameters/index.js
@@ -21,8 +21,8 @@ module.exports = (ctx) => ({
       return ctx.event.request[name];
     }
 
-    const slot = ctx.event.request.intent.slots[camelCaseToScreamingSnake(name)] || {};
-    const id = _.get(slot, ['resolutions', 'resolutionsPerAuthority', 0, 'values', 0, 'value', 'id']);
-    return id || slot.value;
+    const intent = ctx.event.request.intent || {};
+    const id = _.get(intent, ['slots', camelCaseToScreamingSnake(name), 'resolutions', 'resolutionsPerAuthority', 0, 'values', 0, 'value', 'id']);
+    return id || _.get(intent, ['slots', camelCaseToScreamingSnake(name), 'value']);
   },
 });
